requirements.txt
torch 
transformers 
tokenizers 
datasets 
biopython 
peft 
accelerate 
bitsandbytes
